+++ src/app/app.module.ts
@@ -1,24 +1,28 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
+/* modules */
 import { AppRoutingModule } from './app-routing.module';
+import { ProductModule } from './product/product.module';
+import { CartModule } from './cart/cart.module';
+import { OrderModule } from './order/order.module';
+import { SharedModule } from './shared/shared.module';
+
+/* components */
 import { AppComponent } from './app.component';
-import { FirstComponent } from './first/first.component';
-import { ProductComponent } from './product/product.component';
-import { ProductListComponent } from './product/product-list/product-list.component';
-import { CartComponent } from './cart/cart.component';
+import { HoverElementDirective } from './shared/hover-element.directive';
 
 @NgModule({
   declarations: [
-    AppComponent,
-    FirstComponent,
-    ProductComponent,
-    ProductListComponent,
-    CartComponent
+    AppComponent
   ],
   imports: [
     BrowserModule,
-    AppRoutingModule
+    AppRoutingModule,
+    ProductModule,
+    CartModule,
+    OrderModule,
+    SharedModule
   ],
   providers: [],
   bootstrap: [AppComponent]