+++ src/app/cart/cart-item/cart-item.component.ts
@@ -1,0 +1,32 @@
+import { Component, OnInit, Input } from '@angular/core';
+
+import { CartService } from '../cart.service';
+
+@Component({
+  selector: 'app-cart',
+  templateUrl: 'cart-item.component.html',
+  styleUrls: ['cart-item.component.scss']
+})
+export class CartItemComponent implements OnInit {
+
+  constructor(private cartService: CartService) { }
+
+  @Input() product: any;
+
+  ngOnInit() {
+  }
+
+  onRemove(product) {
+    this.cartService.removeProduct(product);
+
+  }
+
+  onIncrease(product) {
+    this.cartService.changeProductQuantity(product, 'sum');
+  }
+
+  onDecrease(product) {
+    this.cartService.changeProductQuantity(product, 'sub');
+  }
+
+}