+++ src/app/product/product-list/product-list.component.ts
@@ -1,6 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, EventEmitter, Input, Output } from '@angular/core';
 
 import { ProductService } from '../product.service';
+import { CartService } from '../../cart/cart.service';
+import { Product } from '../product.model';
 
 @Component({
   selector: 'app-product-list',
@@ -12,7 +14,14 @@ export class ProductListComponent implements OnInit {
 
   products;
 
-  constructor(public productService: ProductService) { }
+  constructor(
+    public productService: ProductService,
+    private cartService: CartService
+  ) { }
+
+  onBuyProduct(product: Product): void {
+    this.cartService.addToCart(product);
+  }
 
   ngOnInit() {
     this.products = this.productService.getProducts();