+++ src/app/shared/hover-element.directive.ts
@@ -1,0 +1,20 @@
+import { Directive, HostBinding, HostListener } from '@angular/core';
+
+@Directive({
+  selector: '[appHover]'
+})
+export class HoverElementDirective {
+
+  @HostBinding('style.background') private background;
+
+  @HostListener('mouseenter', ['$event'])
+  enter(event: Event) {
+    this.background = 'lightgreen';
+  }
+
+  @HostListener('mouseleave', ['$event'])
+  leave(event: Event) {
+    this.background = 'transparent';
+  }
+
+}