+++ src/app/shared/shared.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { HoverElementDirective } from './hover-element.directive';
+
+@NgModule({
+  declarations: [HoverElementDirective],
+  exports: [HoverElementDirective],
+  imports: [
+    CommonModule
+  ]
+})
+export class SharedModule { }